Fixed unit switch so that math rounding doesn't change original displayed temperature/wind speed values
Originally, unit switching did the math on the rounded value stored in the HTML of each temperature/wind element.  This meant that if the unit was switched multiple times, the rounding each time could result in a different value being displayed than was originally pulled from the API.  Now pushing the API data to an array so that all unit switching math is done on the original data.